Move Colour to utils and remove feature guards
