+++ command_attr/src/structures.rs
@@ -421,7 +421,7 @@ impl Colour {
 impl ToTokens for Colour {
     fn to_tokens(&self, stream: &mut TokenStream2) {
         let value = self.0;
-        let path = quote!(serenity::utils::Colour);
+        let path = quote!(serenity::model::colour::Colour);
 
         stream.extend(quote! {
             #path(#value)
