+++ src/framework/standard/structures/mod.rs
@@ -7,9 +7,9 @@ use futures::future::BoxFuture;
 use super::Args;
 use crate::client::Context;
 use crate::model::channel::Message;
+use crate::model::colour::Colour;
 use crate::model::id::UserId;
 use crate::model::permissions::Permissions;
-use crate::utils::Colour;
 
 pub mod buckets;
 mod check;
